move ensure_tuple to utils
