+++ pydantic_collab/_utils.py
@@ -19,7 +19,7 @@ from pydantic_ai import (
 )
 
 if TYPE_CHECKING:
-    from ._types import HandoffData, PromptBuilderContext
+    from ._types import HandoffData, PromptBuilderContext, T
 
 # =============================================================================
 # Message History Utilities
@@ -266,3 +266,23 @@ def default_build_agent_prompt(ctx: PromptBuilderContext) -> str:
     output_str = 's\n'.join(output_instructions)
 
     return output_str
+
+
+# For some reason tuple[Unpack[T]]  doesn't work in ruff
+def ensure_tuple(value: T) -> tuple[T] | T | None:
+    """Convert a value to a tuple, handling various input types.
+
+    Args:
+        value: Value to convert - can be None, tuple, list, set, frozenset, or single item
+
+    Returns:
+        None if value is None, the original tuple if already a tuple,
+        a tuple of the items if a collection, or a single-item tuple for other values
+    """
+    if value is None:
+        return None
+    if isinstance(value, tuple):
+        return value
+    elif isinstance(value, (list, set, frozenset)):
+        return tuple(value)
+    return (value,)
